Refactor JCloudsBuildWrapper
