+++ src/main/java/jenkins/plugins/openstack/compute/JCloudsBuildWrapper.java
@@ -1,6 +1,12 @@
 package jenkins.plugins.openstack.compute;
 
+import com.google.common.collect.ImmutableList;
+import com.google.common.util.concurrent.FutureCallback;
+import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import com.google.common.util.concurrent.ListeningExecutorService;
 import hudson.Extension;
+import hudson.Functions;
 import hudson.Launcher;
 import hudson.Util;
 import hudson.model.AbstractBuild;
@@ -14,22 +20,21 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.Callable;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import javax.annotation.Nonnull;
 
-import jenkins.plugins.openstack.compute.internal.NodePlan;
+import jenkins.plugins.openstack.compute.internal.DestroyMachine;
 import jenkins.plugins.openstack.compute.internal.Openstack;
-import jenkins.plugins.openstack.compute.internal.ProvisionPlannedInstancesAndDestroyAllOnError;
-import jenkins.plugins.openstack.compute.internal.RunningNode;
-import jenkins.plugins.openstack.compute.internal.TerminateNodes;
 
+import org.jenkinsci.plugins.resourcedisposer.AsyncResourceDisposer;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.openstack4j.model.compute.Server;
 
 import com.google.common.base.Function;
-import com.google.common.base.Supplier;
 import com.google.common.collect.Iterables;
 import com.google.common.util.concurrent.MoreExecutors;
 
@@ -46,34 +51,81 @@ public class JCloudsBuildWrapper extends BuildWrapper {
         return instancesToRun;
     }
 
-    //
     // convert Jenkins staticy stuff into pojos; performing as little critical stuff here as
     // possible, as this method is very hard to test due to static usage, etc.
-    //
     @Override
     public Environment setUp(final AbstractBuild build, Launcher launcher, final BuildListener listener) {
+        // TODO check quota here to abort the task if we do not have enough resources
 
         // eagerly lookup node supplier so that errors occur before we attempt to provision things
         Iterable<NodePlan> nodePlans = Iterables.transform(instancesToRun, new Function<InstancesToRun, NodePlan>() {
 
             @SuppressWarnings("unchecked")
             public NodePlan apply(InstancesToRun instance) {
-                String cloudName = instance.cloudName;
+                JCloudsCloud cloud = JCloudsCloud.getByName(instance.cloudName);
                 String templateName = Util.replaceMacro(instance.getActualTemplateName(), build.getBuildVariableResolver());
-                JCloudsCloud cloud = JCloudsCloud.getByName(cloudName);
                 JCloudsSlaveTemplate template = cloud.getTemplate(templateName);
                 if (template == null) throw new IllegalArgumentException("No such template " + templateName);
-                Supplier<Server> nodeSupplier = new ServerSupplier(cloud, template);
-                return new NodePlan(cloudName, templateName, instance.count, nodeSupplier);
+                return new NodePlan(cloud, template, instance.count);
             }
         });
 
-        final TerminateNodes terminateNodes = new TerminateNodes();
+        ListeningExecutorService executor = MoreExecutors.listeningDecorator(Computer.threadPoolForRemoting);
+        final ImmutableList.Builder<RunningNode> cloudTemplateNodeBuilder = ImmutableList.builder();
 
-        ProvisionPlannedInstancesAndDestroyAllOnError provisioner = new ProvisionPlannedInstancesAndDestroyAllOnError(
-                MoreExecutors.listeningDecorator(Computer.threadPoolForRemoting), listener, terminateNodes);
+        final ImmutableList.Builder<ListenableFuture<Server>> plannedInstancesBuilder = ImmutableList.builder();
 
-        final Iterable<RunningNode> runningNode = provisioner.apply(nodePlans);
+        final AtomicInteger failedLaunches = new AtomicInteger();
+
+        for (final NodePlan nodePlan : nodePlans) {
+            for (int i = 0; i < nodePlan.getCount(); i++) {
+                final int index = i;
+                listener.getLogger().printf(
+                        "Queuing cloud instance: #%d %s %s%n",
+                        index, nodePlan.getCloud(), nodePlan.getTemplate()
+                );
+
+                ListenableFuture<Server> provisionTemplate = executor.submit(nodePlan.getNodeSupplier());
+
+                Futures.addCallback(provisionTemplate, new FutureCallback<Server>() {
+                    public void onSuccess(Server result) {
+                        if (result != null) {
+                            synchronized (cloudTemplateNodeBuilder) {
+                                // Builder in not threadsafe
+                                cloudTemplateNodeBuilder.add(new RunningNode(nodePlan.getCloud(), result));
+                            }
+                        } else {
+                            failedLaunches.incrementAndGet();
+                        }
+                    }
+
+                    public void onFailure(@Nonnull Throwable t) {
+                        failedLaunches.incrementAndGet();
+                        listener.error(
+                                "Error while launching instance: #%d, %s %s:%n%s%n",
+                                index, nodePlan.getCloud(), nodePlan.getTemplate(), Functions.printThrowable(t)
+                        );
+                    }
+                });
+
+                plannedInstancesBuilder.add(provisionTemplate);
+            }
+        }
+
+        // block until all complete
+        List<Server> nodesActuallyLaunched = Futures.getUnchecked(Futures.successfulAsList(plannedInstancesBuilder.build()));
+
+        final ImmutableList<RunningNode> runningNode = cloudTemplateNodeBuilder.build();
+
+        if (failedLaunches.get() > 0) {
+            terminateNodes(runningNode);
+            throw new IllegalStateException("One or more instances failed to launch.");
+        }
+
+        assert runningNode.size() == nodesActuallyLaunched.size() : String.format(
+                "expected nodes from callbacks to be the same count as those from the list of futures!%n" + "fromCallbacks:%s%nfromFutures%s%n",
+                runningNode, nodesActuallyLaunched);
+
 
         final String ipsString = getIpsString(runningNode);
         return new Environment() {
@@ -84,7 +136,7 @@ public class JCloudsBuildWrapper extends BuildWrapper {
 
             @Override
             public boolean tearDown(AbstractBuild build, final BuildListener listener) throws IOException, InterruptedException {
-                terminateNodes.apply(runningNode);
+                terminateNodes(runningNode);
                 return true;
             }
         };
@@ -106,18 +158,66 @@ public class JCloudsBuildWrapper extends BuildWrapper {
         return Util.join(ips, ",");
     }
 
-    private static final class ServerSupplier implements Supplier<Server> {
+    private static void terminateNodes(Iterable<RunningNode> runningNodes) {
+        AsyncResourceDisposer disposer = AsyncResourceDisposer.get();
+        for (RunningNode rn: runningNodes) {
+            disposer.dispose(new DestroyMachine(rn.getCloudName(), rn.getNode().getId()));
+        }
+    }
 
-        private final @Nonnull JCloudsCloud cloud;
-        private final @Nonnull JCloudsSlaveTemplate template;
+    public static class RunningNode {
+        private final String cloud;
+        private final Server node;
 
-        private ServerSupplier(@Nonnull JCloudsCloud cloud, @Nonnull JCloudsSlaveTemplate template) {
+        RunningNode(String cloud, Server node) {
             this.cloud = cloud;
-            this.template = template;
+            this.node = node;
         }
 
-        @Override public Server get() {
-            return template.provision(cloud);
+        public String getCloudName() {
+            return cloud;
+        }
+
+        public Server getNode() {
+            return node;
+        }
+    }
+
+    public static class NodePlan {
+        private final JCloudsCloud cloud;
+        private final JCloudsSlaveTemplate template;
+        private final int count;
+
+        NodePlan(JCloudsCloud cloud, JCloudsSlaveTemplate template, int count) {
+            this.cloud = cloud;
+            this.template = template;
+            this.count = count;
+        }
+
+        public String getCloud() {
+            return cloud.name;
+        }
+
+        public String getTemplate() {
+            return template.name;
+        }
+
+        public int getCount() {
+            return count;
+        }
+
+        Callable<Server> getNodeSupplier() {
+            final JCloudsCloud cloud1 = cloud;
+            final JCloudsSlaveTemplate template1 = template;
+            return new Callable<Server>() {
+                private final @Nonnull JCloudsCloud cloud = cloud1;
+                private final @Nonnull JCloudsSlaveTemplate template = template1;
+
+                @Override
+                public Server call() throws Exception {
+                    return template.provision(cloud);
+                }
+            };
         }
     }
